refactor(entry): extract list/continuation filter builders into filters.rs
entry.rs was ~2330 lines. Move the five private WHERE-clause / index-hint
builders (is_no_entry_side_predicate, published_sort_entry_hint,
apply_filter_conditions, apply_time_conditions, apply_continuation_condition)
into entry/filters.rs as pub(super) helpers. Types stay in mod.rs; call sites
are unchanged via a use re-export.
